+++ daskms/query.py
@@ -1,9 +1,5 @@
 # -*- coding: utf-8 -*-
 
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-
 
 def select_clause(select_cols):
     if select_cols is None or len(select_cols) == 0: