+++ daskms/tests/test_optional.py
@@ -9,10 +9,6 @@ Pass the --optional command line argument to py.test to enable them
 $ py.test --optional
 """
 
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-
 import os
 from pprint import pprint  # noqa